+++ modules/canvaskit/canvaskit/types/canvaskit-wasm-tests.ts
@@ -237,6 +237,13 @@ function imageTests(CK: CanvasKit, imgElement?: HTMLImageElement) {
     const buff = new ArrayBuffer(10);
     const img = CK.MakeImageFromEncoded(buff); // $ExpectType Image | null
     const img2 = CK.MakeImageFromCanvasImageSource(imgElement); // $ExpectType Image
+    const img3 = CK.MakeImage({ // $ExpectType Image | null
+      width: 1,
+      height: 1,
+      alphaType: CK.AlphaType.Premul,
+      colorType: CK.ColorType.RGBA_8888,
+      colorSpace: CK.ColorSpace.SRGB
+    }, Uint8Array.of(255, 0, 0, 250), 4);
     if (!img) return;
     const dOne = img.encodeToData(); // $ExpectType Data
     const dTwo = img.encodeToDataWithFormat(CK.ImageFormat.JPEG, 97);
@@ -244,7 +251,7 @@ function imageTests(CK: CanvasKit, imgElement?: HTMLImageElement) {
     const h = img.height();
     const w = img.width();
     const shader = img.makeShader(CK.TileMode.Decal, CK.TileMode.Repeat); // $ExpectType Shader
-    const pixels = img.readPixels(85, 1000, {// $Uint8Array | Float32Array | null
+    const pixels = img.readPixels(85, 1000, { // $ExpectType Float32Array | Uint8Array | null
         width: 79,
         height: 205,
         colorType: CK.ColorType.RGBA_8888,
@@ -259,6 +266,9 @@ function imageTests(CK: CanvasKit, imgElement?: HTMLImageElement) {
         alphaType: CK.AlphaType.Unpremul,
         colorSpace: CK.ColorSpace.SRGB,
     }, m, 4 * 85);
+    const ii = img.getImageInfo(); // $ExpectType PartialImageInfo
+    const cs = img.getColorSpace(); // $ExpectType ColorSpace
+    cs.delete();
     img.delete();
 }