+++ modules/canvaskit/CHANGELOG.md
@@ -6,9 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - `getImageInfo()` and `getColorSpace()` to the `Image` type.
+
+### Changed
+ - `MakeImage` is now documented in the Typescript types (index.d.ts). The parameters have been
+   streamlined to align with other, similar APIs.
+
 ## [0.20.0] - 2020-11-12
 
-## Added
+### Added
  - `MakeFractalNoise`, `MakeImprovedNoise`, and `MakeTurbulence` have been added to
    `CanvasKit.Shader`.
  - `MakeRasterDirectSurface` for giving the user direct access to drawn pixels.